@@ -1,7 +1,6 @@
 const { interpolate } = require('@usebruno/common');
 const { each, forOwn, cloneDeep, find } = require('lodash');
 const FormData = require('form-data');
-const { mockDataFunctions } = require('./faker-functions');
 
 const getContentType = (headers = {}) => {
   let contentType = '';
@@ -12,28 +11,6 @@ const getContentType = (headers = {}) => {
   });
 
   return contentType;
-};
-
-const interpolateMockVars = (str, { escapeJSONStrings }) => {
-  const patternRegex = /\{\{\$(\w+)\}\}/g;
-  return str.replace(patternRegex, (match, keyword) => {
-    let replacement = mockDataFunctions[keyword]?.();
-
-    if (replacement === undefined) return match;
-    replacement = String(replacement);
-
-    if (!escapeJSONStrings) return replacement;
-    // All the below chars inside of a JSON String field
-    // will make it invalid JSON. So we will have to escape them with `\`.
-    // This is not exhaustive but selective to what faker-js can output.
-    if (!/[\\\n\r\t\"]/.test(replacement)) return replacement;
-    return replacement
-      .replace(/\\/g, '\\\\')
-      .replace(/\n/g, '\\n')
-      .replace(/\r/g, '\\r')
-      .replace(/\t/g, '\\t')
-      .replace(/\"/g, '\\"');
-  });
 };
 
 const interpolateVars = (request, envVariables = {}, runtimeVariables = {}, processEnvVars = {}) => {
@@ -78,7 +55,9 @@ const interpolateVars = (request, envVariables = {}, runtimeVariables = {}, proc
       }
     };
 
-    return interpolateMockVars(interpolate(str, combinedVars), { escapeJSONStrings });
+    return interpolate(str, combinedVars, {
+      escapeJSONStrings
+    });
   };
 
   request.url = _interpolate(request.url);
@@ -97,12 +76,16 @@ const interpolateVars = (request, envVariables = {}, runtimeVariables = {}, proc
   if (contentType.includes('json') && !Buffer.isBuffer(request.data)) {
     if (typeof request.data === 'string') {
       if (request.data.length) {
-        request.data = _interpolate(request.data, { escapeJSONStrings: true });
+        request.data = _interpolate(request.data, {
+          escapeJSONStrings: true
+        });
       }
     } else if (typeof request.data === 'object') {
       try {
         const jsonDoc = JSON.stringify(request.data);
-        const parsed = _interpolate(jsonDoc, { escapeJSONStrings: true });
+        const parsed = _interpolate(jsonDoc, {
+          escapeJSONStrings: true
+        });
         request.data = JSON.parse(parsed);
       } catch (err) {}
     }
